Get midengine building with emscripten
Command:

  emcc -o benchmark.js -O3 -s ALLOW_MEMORY_GROWTH=1 -std=c++1z -Wall -Werror -I. -Ibazel-genfiles -Ibazel-pentago/external/{boost,tinyformat,com_google_googletest/googletest/include} pentago/mid/{benchmark,halfsuper,midengine}.cc pentago/high/board.cc bazel-bin/pentago/base/gen/tables.cc pentago/utility/{debug,exceptions,log}.cc pentago/base/{board,count,superscore,symmetry}.cc

Results:

  % time node benchmark.js
  time = 11.263 s

  real	0m12.034s
  user	0m13.689s
  sys	0m0.274s

So a bit more than 1 s overhead starting from scratch, which seems fine.

+++ pentago/utility/array.h
@@ -302,8 +302,8 @@ public:
   RawArray() : Base(Vector<int,d>(), 0) {}
   RawArray(int size, T* data) : Base(vec(size), data) {}
   RawArray(Vector<int,d> shape, T* data) : Base(shape, data) {}
-  RawArray(vector<value_type>& x) : RawArray(CHECK_CAST_INT(x.size()), x.data()) {}
-  RawArray(const vector<value_type>& x) : RawArray(CHECK_CAST_INT(x.size()), x.data()) {}
+  RawArray(vector<value_type>& x) : RawArray(CHECK_CAST_INT(uint64_t(x.size())), x.data()) {}
+  RawArray(const vector<value_type>& x) : RawArray(CHECK_CAST_INT(uint64_t(x.size())), x.data()) {}
 
   template<class S> RawArray(const Array<S,d>& array)
     : Base(array.shape(), array.data()) {}
